requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
glfw==2.5.5
idna==3.4
numpy==1.23.5
Pillow==9.3.0
PyGLM==2.6.0
PyOpenGL==3.1.6
requests==2.28.1
urllib3==1.26.13
